+++ src/App.vue
@@ -84,5 +84,4 @@ $navbar-breakpoint: $tablet;
 $dimensions: 16 24 32 48 64 96 128 256;
 
 @import "~bulma";
-@import "~buefy/src/scss/buefy";
 </style>